Move ParentsMergedState to merging folder (fix file name)
